90% System Payment

+++ backend/entity/tier.go
@@ -5,7 +5,7 @@ import "gorm.io/gorm"
 type Tier struct {	
 	gorm.Model
 	TierName	 string 
-	Price 		 int
+	Price 		 float32
 	
 	Users []User `gorm:"foreignKey:TierID"`
 }
